+++ crates/ndc-sqlserver/tests/schema_tests.rs
@@ -2,6 +2,9 @@ pub mod common;
 
 #[tokio::test]
 async fn get_schema() {
-    let result = common::helpers::get_schema(common::helpers::create_router().await).await;
+    let result = common::helpers::get_schema(
+        common::helpers::create_router(common::helpers::SQLSERVER_CONNECTION_STRING).await,
+    )
+    .await;
     insta::assert_json_snapshot!(result);
 }